Create a common component directory and move Button/ into it
