+++ src/components/Contact/form.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import NetlifyForm from './netlifyForm';
 import Recaptcha from 'react-google-recaptcha';
-import Button, { ButtonThemes, ButtonSizes } from '../Button';
+import Button, { ButtonThemes, ButtonSizes } from '../common/Button';
 
 const RECAPTCHA_KEY = process.env.GATSBY_SITE_RECAPTCHA_KEY;
 const RECAPTCHA_NETLIFY_PROP = 'g-recaptcha-response';
